Update tests structure and handles intervened values
