+++ tests/test_inference_queries.py
@@ -16,8 +16,6 @@ logging.disable(logging.INFO)
 from causal_reasoning.causal_model import CausalModel
 from causal_reasoning.utils._enum import DataExamplesPaths
 
-from auxiliary import genGraph, true_value
-
 
 class TestInferenceAlgorithm(unittest.TestCase):
     def test_binary_balke_pearl_example(self):
@@ -215,209 +213,6 @@ class TestInferenceAlgorithm(unittest.TestCase):
                 ["MS-A_Latency"], ["DB_Change"], ["DB_Latency"]
             )
         )
-
-        # model_2.set_interventions([(intervention_1, 0)])
-        # model_2.set_target((target, 0))
-        # print(f"{intervention_1}: PN = {model_2.inference_intervention_query()}")
-
-        # model_2.set_interventions([(intervention_1, 1)])
-        # model_2.set_target((target, 1))
-        # print(f"{intervention_1}: PS = {model_2.inference_intervention_query()}")
-
-        # model_2.set_interventions([(intervention_2, 0)])
-        # model_2.set_target((target, 0))
-        # print(f"{intervention_2}: PN = {model_2.inference_intervention_query()}")
-
-        # model_2.set_interventions([(intervention_2, 1)])
-        # model_2.set_target((target, 1))
-        # print(f"{intervention_2}: PS = {model_2.inference_intervention_query()}")
-
-
-class TestIdentifiableInterventionQueries(unittest.TestCase):
-    def test_identifiable_queries_via_subtests(self):
-        cases = [
-            (1, 1, DataExamplesPaths.CSV_N1M1),
-            (2, 1, DataExamplesPaths.CSV_N2M1),
-        ]
-        unobs = ["U1", "U2", "U3"]
-        interventions = [(0, 0), (0, 1), (1, 0), (1, 1)]
-
-        for N, M, csv_example in cases:
-            with self.subTest(N=N, M=M):
-                edges = genGraph(N=N, M=M)
-                csv_path = os.path.join(PROJECT_ROOT, csv_example.value)
-                df = pd.read_csv(csv_path)
-                df["U3"] = df["Y"].copy()
-
-                model = CausalModel(
-                    data=df,
-                    edges=edges,
-                    unobservables_labels=unobs,
-                )
-
-                for target_value, intervention_value in interventions:
-                    with self.subTest(
-                        N=N, M=M, target=target_value, intervention=intervention_value
-                    ):
-                        model.set_interventions([("X", intervention_value)])
-                        model.set_target(("Y", target_value))
-
-                        identifiable_value = model.identifiable_intervention_query()
-
-                        tv = true_value(N, M, target_value, intervention_value, df)
-
-                        self.assertAlmostEqual(
-                            identifiable_value,
-                            tv,
-                            places=2,
-                            msg=f"Values do not match for N={N},M={M},Y={target_value},do(X={intervention_value})",
-                        )
-
-
-class TestMNCases(unittest.TestCase):
-    def test_intervention_queries_via_subtests(self):
-        cases = [
-            (1, 1, DataExamplesPaths.CSV_N1M1),
-            (2, 1, DataExamplesPaths.CSV_N2M1),
-            # Too long to run
-            # (3, 1, DataExamplesPaths.CSV_N3M1),
-            # (4, 1, DataExamplesPaths.CSV_N4M1),
-            # (1, 2, DataExamplesPaths.CSV_N1M2),
-        ]
-        unobs = ["U1", "U2", "U3"]
-        interventions = [(0, 0), (0, 1), (1, 0), (1, 1)]
-
-        for N, M, csv_example in cases:
-            with self.subTest(N=N, M=M):
-                edges = genGraph(N=N, M=M)
-                csv_path = os.path.join(PROJECT_ROOT, csv_example.value)
-                df = pd.read_csv(csv_path)
-                df["U3"] = df["Y"].copy()
-
-                model = CausalModel(
-                    data=df,
-                    edges=edges,
-                    unobservables_labels=unobs,
-                )
-
-                self.assertFalse(
-                    model.are_d_separated_in_complete_graph(["X"], ["Y"], unobs),
-                    msg=f"d-separation failed for N={N}, M={M}",
-                )
-
-                for target_value, intervention_value in interventions:
-                    with self.subTest(
-                        N=N, M=M, target=target_value, intervention=intervention_value
-                    ):
-                        model.set_interventions([("X", intervention_value)])
-                        model.set_target(("Y", target_value))
-                        lower, upper = model.inference_intervention_query()
-
-                        tv = model.identifiable_intervention_query()
-
-                        self.assertGreaterEqual(
-                            float(upper),
-                            tv,
-                            msg=f"upper bound too low for N={N},M={M},Y={target_value},do(X={intervention_value})",
-                        )
-                        self.assertLessEqual(
-                            float(lower),
-                            tv,
-                            msg=f"lower bound too high for N={N},M={M},Y={target_value},do(X={intervention_value})",
-                        )
-
-class TestIsIdentifiableIntervention(unittest.TestCase):
-    def setUp(self):
-        self.df_xy = pd.DataFrame({
-            'X': [0, 1] * 10,
-            'Y': [1, 0] * 10
-        })
-        self.df_xwy = pd.DataFrame({
-            'X': [0, 1] * 10,
-            'W': [0, 1] * 10,
-            'Y': [1, 0] * 10
-        })
-        self.df_xzy = pd.DataFrame({
-            'X': [0, 1] * 10,
-            'Z': [0, 1] * 10,
-            'Y': [1, 0] * 10
-        })
-
-    def test_backdoor_non_identifiable(self):
-        # Pure confounder U -> X, U -> Y
-        graph = "U -> X, U -> Y, X -> Y"
-        model = CausalModel(
-            data=self.df_xy,
-            edges=graph,
-            unobservables_labels=['U']
-        )
-        identifiable, method, detail = model.is_identifiable_intervention(
-            interventions=[('X', 0)],
-            target=('Y', 1)
-        )
-        self.assertFalse(identifiable)
-        self.assertEqual(method, "unobservable_confounding")
-        self.assertIsNone(detail)
-
-    def test_frontdoor_identifiable(self):
-        # Front-door: U1 -> X,Y; X -> W -> Y; U2 -> W (harmless latent on W)
-        graph = "U1 -> X, U1 -> Y, X -> W, W -> Y, U2 -> W"
-        model = CausalModel(
-            data=self.df_xwy,
-            edges=graph,
-            unobservables_labels=['U1', 'U2']
-        )
-        identifiable, method, detail = model.is_identifiable_intervention(
-            interventions=[('X', 0)],
-            target=('Y', 1)
-        )
-        self.assertTrue(identifiable)
-        self.assertEqual(method, "frontdoor")
-        # The minimal observed front-door set should be {'W'}
-        self.assertEqual(detail, frozenset({'W'}))
-
-    def test_iv_identifiable(self):
-        # IV: U1 -> X,Y; U2 -> Z; Z -> X -> Y
-        graph = "U1 -> X, U1 -> Y, U2 -> Z, Z -> X, X -> Y"
-        model = CausalModel(
-            data=self.df_xzy,
-            edges=graph,
-            unobservables_labels=['U1', 'U2']
-        )
-        identifiable, method, detail = model.is_identifiable_intervention(
-            interventions=[('X', 0)],
-            target=('Y', 1)
-        )
-        self.assertTrue(identifiable)
-        self.assertEqual(method, "instrumental_variable")
-        # The identified instrument should be 'Z'
-        self.assertEqual(detail, 'Z')
-
-    def test_missing_intervention_raises(self):
-        graph = "U -> X, U -> Y, X -> Y"
-        model = CausalModel(
-            data=self.df_xy,
-            edges=graph,
-            unobservables_labels=['U']
-        )
-        with self.assertRaises(Exception):
-            model.is_identifiable_intervention(
-                interventions=[],
-                target=('Y', 1)
-            )
-
-    def test_missing_target_raises(self):
-        graph = "U -> X, U -> Y, X -> Y"
-        model = CausalModel(
-            data=self.df_xy,
-            edges=graph,
-            unobservables_labels=['U']
-        )
-        with self.assertRaises(Exception):
-            model.is_identifiable_intervention(
-                interventions=[('X', 0)],
-                target=None
-            )
 
 
 if __name__ == "__main__":
